mruby: Move grn_accessor related codes in a new file
