+++ lib/mrb/sources.am
@@ -1,4 +1,6 @@
 libgrnmrb_la_SOURCES =				\
+	mrb_accessor.c				\
+	mrb_accessor.h				\
 	mrb_expr.c				\
 	mrb_expr.h				\
 	mrb_obj.c				\
